refactor:refactor describe,explain and drop
