+++ src/sql/src/parsers.rs
@@ -16,6 +16,9 @@ mod alter_parser;
 pub(crate) mod copy_parser;
 pub(crate) mod create_parser;
 pub(crate) mod delete_parser;
+pub(crate) mod describe_parser;
+pub(crate) mod drop_parser;
+pub(crate) mod explain_parser;
 pub(crate) mod insert_parser;
 pub(crate) mod query_parser;
 pub(crate) mod show_parser;
